requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
transformers==4.36.2
espnet==0.10.6
Pillow==10.0.0
datasets==2.20.0
matplotlib==3.8.0
seaborn==0.13.0
jiwer==3.0.4
evaluate==0.4.3
sentencepiece==0.2.0